minor fixe

+++ src/main/java/bookstore/controller/ShoppingCartController.java
@@ -2,6 +2,7 @@ package bookstore.controller;
 
 import bookstore.dto.cart.AddToCartDto;
 import bookstore.dto.cart.CartResponseDto;
+import bookstore.dto.cartitem.CartItemFullInfoResponseDto;
 import bookstore.dto.cartitem.CartItemResponseDto;
 import bookstore.model.User;
 import bookstore.service.ShoppingCartService;
@@ -15,6 +16,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -42,8 +44,8 @@ public class ShoppingCartController {
     @PostMapping()
     @ApiResponse(responseCode = "204", description = "Item added to the cart")
     public void addCartItemTo(Authentication authentication,
-            @RequestBody @Parameter(description = "Item to add to the cart")
-            @Valid AddToCartDto addToCartDto) {
+                              @RequestBody @Parameter(description = "Item to add to the cart")
+                              @Valid AddToCartDto addToCartDto) {
         service.addItemToCart(addToCartDto, (User) authentication.getPrincipal());
     }
 
@@ -56,11 +58,18 @@ public class ShoppingCartController {
                             + " \"book\": {\"id\": 1, \"title\":"
                             + " \"Book Title\"}, \"quantity\": 2}"))
     })
-    @ApiResponse(responseCode = "404", description = "Cart item not found")
-    public CartItemResponseDto updateCartItem(
-            @PathVariable @Parameter(description = "ID of the cart item to update") Long id,
-            @RequestBody @Parameter(description = "Updated cart item details")
-            @Valid AddToCartDto dto) {
+    @ApiResponse(responseCode = "404", description = "Cart item not found or doesn't"
+            + " belong to the user's cart")
+    public CartItemResponseDto updateCartItem(Authentication authentication,
+                                              @PathVariable
+                                              @Parameter
+                                                      (description = "ID of the cart"
+                                                              + " item to update") Long id,
+                                              @RequestBody
+                                                  @Parameter(description =
+                                                          "Updated cart item details")
+                                              @Valid AddToCartDto dto) {
+        checkCartItemModificationPermission(authentication, id);
         return service.updateCartItem(id, dto);
     }
 
@@ -70,8 +79,25 @@ public class ShoppingCartController {
             @ApiResponse(responseCode = "204", description = "Cart item deleted"),
             @ApiResponse(responseCode = "404", description = "Cart item not found")
     })
-    public void deleteCartItem(@PathVariable @Parameter
+    public void deleteCartItem(Authentication authentication, @PathVariable @Parameter
             (description = "ID of the cart item to delete") Long id) {
+        checkCartItemModificationPermission(authentication, id);
+
         service.deleteCartItem(id);
+    }
+
+    private void checkCartItemModificationPermission(Authentication authentication,
+                                                     @Parameter
+                                                             (description = "ID of the cart"
+                                                                     + " item to delete")
+                                                     @PathVariable Long id) {
+        CartItemFullInfoResponseDto cartItem = service.getCartItemById(id);
+
+        User user = (User) authentication.getPrincipal();
+        CartResponseDto cart = service.getShoppingCartByUser(user);
+        if (cart == null || !cartItem.getShoppingCartId().equals(cart.getId())) {
+            throw new AccessDeniedException("this shopping cart item doesn't"
+                    + " belong to the current user");
+        }
     }
 }
